+++ packages/core/src/utils.js
@@ -1,2 +1,4 @@
 // @flow
 export let isBrowser = typeof document !== 'undefined'
+
+export const hasOwnProperty = Object.prototype.hasOwnProperty
